requirements.txt
pywebview
numpy~=1.21.6
scikit-learn~=1.0.2
tensorflow==2.11.0
pillow~=9.5.0